move MediaUri into a media folder
